ENH: Add a helper function to create wrongly-defaced augmentations
Creates augmentations by flipping the face mask along the AP and/or
the IS axes.

A simple heuristic is used to determine (and later drop) the left-right
axes (LR). First, a bounding box of masked voxels is estimated. Then,
the AP axes is assumed to be that containing less voxels in the bounding
box and the IS is assumed to be that containing most voxels.

Then the combinations of flipped masks and data are estimated.

Results are stored into a temporary folder.