ENH: Add a helper function to create wrongly-defaced augmentations
Creates augmentations by flipping the face mask along the AP and/or
the IS axes.

A simple heuristic is used to determine (and later drop) the left-right
axes (LR). First, a bounding box of masked voxels is estimated. Then,
the AP axes is assumed to be that containing less voxels in the bounding
box and the IS is assumed to be that containing most voxels.

Then the combinations of flipped masks and data are estimated.

Results are stored into a temporary folder.


Former-commit-id: 6e4061c45a99fe162441c6a3a9a81989a2960742
Former-commit-id: b4a02b3e10b6de05f57afd3a8793908fba723935 [formerly 3fb602daedd2ba304b4a488650017fca29048058]
Former-commit-id: 30c7f05a6af94e44537af3ca7ca169317165520d